+++ src/stores/emoji.ts
@@ -1,6 +1,7 @@
 import { defineStore } from "pinia";
 import { EmojiButton } from "@joeattardi/emoji-button";
-import { hideMainUI } from "@/common/funcs";
+import { getNumber, hideMainUI, resetNumber } from "@/common/funcs";
+import "@logseq/libs";
 
 export const useEmojiStore = defineStore("emoji", {
   state: () => ({
@@ -18,7 +19,12 @@ export const useEmojiStore = defineStore("emoji", {
       });
 
       this.picker.on("emoji", async (selection) => {
-        await logseq.Editor.insertAtEditingCursor(selection.emoji);
+        const number = getNumber();
+        resetNumber();
+        const emojis = [...new Array(number)]
+          .map(() => selection.emoji)
+          .join("");
+        await logseq.Editor.insertAtEditingCursor(emojis);
       });
 
       this.picker.on("hidden", async (selection) => {
@@ -31,7 +37,22 @@ export const useEmojiStore = defineStore("emoji", {
 
       return this.picker;
     },
-    async init() {
+    async showPicker() {
+      const { left, top, rect, pos } =
+        await logseq.Editor.getEditingCursorPosition();
+      Object.assign(this.emojiPickerEl.style, {
+        position: "absolute",
+        top: top + rect.top + "px",
+        left: left + rect.left + "px",
+      });
+
+      setTimeout(() => {
+        if (this.picker) {
+          this.picker.showPicker(this.emojiPickerEl);
+        }
+      }, 100);
+    },
+    async initPicker() {
       this.emojiPickerEl = document.createElement("div");
       this.emojiPickerEl.classList.add("emoji-picker-trigger");
       document.getElementById("app").appendChild(this.emojiPickerEl);